+++ news.md
@@ -3,7 +3,10 @@
 * `s3GetFile()`, `s3PutFile()` - functions are changed to read data based on the extention of the file. Fore example you can use it instead of calling `s3GetFile.csv()` if path ends with `.csv`.
 
 * `bqInsertData()` - added fields parameter to force BigQuery types to the given types
-* `disaggregate()` - new function to split data.table from aggregated to individual lines
+
+* `disaggregate()` - new function to split data.table from aggregated to individual lines (@daniel000, #74)
+
+* `ftpRead()`, `ftpListFiles()` - ftp helpers that allow to access CSV files in FTP folders (@byapparov, #75).
 
 # retl 0.1.1